Move config tests to src and scripts
